+++ src/contexts/SettingsProvider.tsx
@@ -8,15 +8,13 @@ interface SidebarProviderProps {
 export const SettingsProvider: React.FC<SidebarProviderProps> = ({
     children,
 }) => {
-    const month = new Date().getMonth() + 1;
+    const month = new Date().getMonth() * -1 + 1;
     const [openedSidebar, setOpenedSidebar] = useState<string>("text");
     const [hide, setHide] = useState<boolean>(false);
     const [hideAnnouncements, setHideAnnouncements] = useState<boolean>(true);
     const [showTutorial, setShowTutorial] = useState<boolean>(false);
     const [showSaveDialog, setShowSaveDialog] = useState<boolean>(false);
-    const [blankCanvas, setBlankCanvas] = useState<boolean>(
-        month === 10 ? true : false
-    );
+    const [blankCanvas, setBlankCanvas] = useState<boolean>(month === 10);
     const [showExperimental, setShowExperimental] = useState<boolean>(false);
     const [openAll, setOpenAll] = useState<boolean>(false);
     const [openTextOption, setOpenTextOption] = useState<string>("name-tag");
@@ -28,6 +26,8 @@ export const SettingsProvider: React.FC<SidebarProviderProps> = ({
     const [allowRefresh, setAllowRefresh] = useState<boolean>(false);
     const [audio, setAudio] = useState<boolean>(false);
     const [loading, setLoading] = useState<number>(0);
+    const [mizuBells, setMizuBells] = useState<boolean>(false);
+    const [settingsLoaded, setSettingsLoaded] = useState<boolean>(false);
 
     useEffect(() => {
         const announcementCookie = localStorage.getItem("5.8.4-announcements");
@@ -42,7 +42,9 @@ export const SettingsProvider: React.FC<SidebarProviderProps> = ({
         if (openAllCookie === "true") {
             setOpenAll(true);
         }
-        const showTutorialCookie = localStorage.getItem("showTutorialAndSetup");
+        const showTutorialCookie = localStorage.getItem(
+            "showTutorialAndSetup-v2"
+        );
         if (!showTutorialCookie || showTutorialCookie === "true") {
             setShowTutorial(true);
         }
@@ -74,11 +76,19 @@ export const SettingsProvider: React.FC<SidebarProviderProps> = ({
         if (storedNameTagInputs) {
             setNameTagInputs(Number(storedNameTagInputs));
         }
+        const mizuBellsWindow = localStorage.getItem("mizu5");
+        if (!mizuBellsWindow || mizuBellsWindow == "true") {
+            setMizuBells(true);
+        }
+        setSettingsLoaded(true);
     }, []);
 
     useEffect(() => {
-        localStorage.setItem("showTutorialAndSetup", String(showTutorial));
+        localStorage.setItem("showTutorialAndSetup-v2", String(showTutorial));
     }, [showTutorial]);
+    useEffect(() => {
+        localStorage.setItem("mizu5", String(mizuBells));
+    }, [mizuBells]);
 
     return (
         <SettingsContext.Provider
@@ -115,6 +125,9 @@ export const SettingsProvider: React.FC<SidebarProviderProps> = ({
                 setAudio,
                 loading,
                 setLoading,
+                mizuBells,
+                setMizuBells,
+                settingsLoaded,
             }}
         >
             {children}