+++ src/utils/i18ninit.tsx
@@ -14,6 +14,7 @@ export const languageNames = {
     es: "Español",
     fr: "Français",
     zh: "简体中文",
+    zhTW: "繁體中文",
     fil: "Filipino",
     ms: "Bahasa Melayu",
     th: "ไทย",