+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+# 5.8.7
+Release date: October 18, 2025
+- Bring back flavor texts
+- Changed countdown to Ena5
+- Fixed settings logic
+
+# 5.8.6
+Release date: October 11, 2025
+- Changed "Withered This Path Of Thorns" date to October 12
+- Changed Y-Offset Initial Setup
+- Added Mizu5 default scene until October 20 04:00 UTC
+
 # 5.8.5
 Release date: October 4, 2025
 - Show guideline when clicking on the canvas